Fix a problem with how delegates were passed through to UIAlertView and SDCAlertView.
We make sure that the SDCAlertViewWrapper’s delegate is passed an instance of SDCAlertViewWrapper;
before this change we were passing the actual alert view.

Includes some fixes to make sure that the instance of SDCAlertViewWrapper is retained by something
during the time that its alert view is visible, since that should be the expectation of consumers of this class.